pyproject.toml
[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"


[project]
name = "layer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.12"
dependencies = [
    "pyyaml>=6.0.3",
    "ruff>=0.15.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
watch = [
    "watchdog>=4.0",
]
dotenv = [
    "python-dotenv>=1.0",
]
aws = [
    "boto3>=1.34",
]
vault = [
    "hvac>=2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
